Change all instances of double to float

+++ examples/1. Basics/readCompass/readCompass.ino
@@ -19,6 +19,6 @@ void setup()
 
 void loop()
 {
-    double yaw = compass.read();
+    float yaw = compass.read();
     Serial.println(yaw);
 }
