+++ src/EVNServo.cpp
@@ -3,7 +3,7 @@
 servo_state_t* EVNServo::servoArgs[];
 bool EVNServo::servos_enabled[] = { false, false, false, false };;
 
-EVNServo::EVNServo(uint8_t port, bool servo_dir, uint16_t range, double start_angle, uint16_t min_pulse_us, uint16_t max_pulse_us, double max_dps)
+EVNServo::EVNServo(uint8_t port, bool servo_dir, uint16_t range, float start_angle, uint16_t min_pulse_us, uint16_t max_pulse_us, float max_dps)
 {
     _servo.servo = new Servo;
     _servo.servo_dir = servo_dir;
@@ -41,33 +41,33 @@ void EVNServo::begin()
     pinMode(_servo.pin, OUTPUT);
     _servo.servo->attach(_servo.pin, 200, 2800);
 
-    double pulse = (double)(_servo.angle / _servo.range) * (double)(_servo.max_pulse_us - _servo.min_pulse_us);
+    float pulse = (float)(_servo.angle / _servo.range) * (float)(_servo.max_pulse_us - _servo.min_pulse_us);
     if (_servo.servo_dir == DIRECT)
-        pulse = (double)_servo.min_pulse_us + pulse;
+        pulse = (float)_servo.min_pulse_us + pulse;
     else
-        pulse = (double)_servo.max_pulse_us - pulse;
+        pulse = (float)_servo.max_pulse_us - pulse;
     _servo.servo->writeMicroseconds(pulse);
 
     attach_servo_interrupt(&_servo);
 }
 
-void EVNServo::write(double angle, uint16_t wait_time_ms, double dps)
+void EVNServo::write(float angle, uint16_t wait_time_ms, float dps)
 {
     this->writeAngle(angle, wait_time_ms, dps);
 }
 
-void EVNServo::writeAngle(double angle, uint16_t wait_time_ms, double dps)
+void EVNServo::writeAngle(float angle, uint16_t wait_time_ms, float dps)
 {
     if (motors_enabled())
     {
         if (dps == 0)
         {
             _servo.angle = constrain(angle, 0, _servo.range);
-            double pulse = (double)(_servo.angle / _servo.range) * (double)(_servo.max_pulse_us - _servo.min_pulse_us);
+            float pulse = (float)(_servo.angle / _servo.range) * (float)(_servo.max_pulse_us - _servo.min_pulse_us);
             if (_servo.servo_dir == DIRECT)
-                pulse = (double)_servo.min_pulse_us + pulse;
+                pulse = (float)_servo.min_pulse_us + pulse;
             else
-                pulse = (double)_servo.max_pulse_us - pulse;
+                pulse = (float)_servo.max_pulse_us - pulse;
             this->writeMicroseconds(pulse);
         }
         else
@@ -102,17 +102,17 @@ void EVNServo::writeMicroseconds(uint16_t pulse_us, uint16_t wait_time_ms)
     }
 }
 
-void EVNServo::writeDutyCycle(double duty_cycle)
+void EVNServo::writeDutyCycle(float duty_cycle)
 {
     if (motors_enabled())
     {
-        double duty_cyclec = constrain(duty_cycle, -1, 1);
-        double pulse = (1 - fabs(duty_cycle)) * 0.5 * (double)(_servo.max_pulse_us - _servo.min_pulse_us);
+        float duty_cyclec = constrain(duty_cycle, -1, 1);
+        float pulse = (1 - fabs(duty_cycle)) * 0.5 * (float)(_servo.max_pulse_us - _servo.min_pulse_us);
 
         if ((_servo.servo_dir == DIRECT) == (duty_cyclec < 0))
-            pulse = (double)_servo.min_pulse_us + pulse;
+            pulse = (float)_servo.min_pulse_us + pulse;
         else
-            pulse = (double)_servo.max_pulse_us - pulse;
+            pulse = (float)_servo.max_pulse_us - pulse;
 
         this->writeMicroseconds(pulse);
         _servo.sweep = false;
