+++ src/EVNServo.h
@@ -14,17 +14,17 @@ typedef struct
 {
 	Servo* servo;
 	volatile bool sweep;
-	volatile double pulse;
+	volatile float pulse;
 	volatile bool servo_dir;
 	volatile uint8_t pin;
 	volatile uint8_t port;
 	volatile uint16_t range;
 	volatile uint16_t min_pulse_us;
 	volatile uint16_t max_pulse_us;
-	volatile double angle;
-	volatile double end_angle;
-	volatile double dps;
-	volatile double max_dps;
+	volatile float angle;
+	volatile float end_angle;
+	volatile float dps;
+	volatile float max_dps;
 	volatile uint64_t last_loop;
 }	servo_state_t;
 
@@ -34,14 +34,14 @@ public:
 
 	static const uint16_t TIMER_INTERVAL_US = 10000;
 
-	EVNServo(uint8_t port, bool servo_dir = DIRECT, uint16_t range = 270, double start_angle = 135, uint16_t min_pulse_us = 600, uint16_t max_pulse_us = 2400, double max_dps = 500);
+	EVNServo(uint8_t port, bool servo_dir = DIRECT, uint16_t range = 270, float start_angle = 135, uint16_t min_pulse_us = 600, uint16_t max_pulse_us = 2400, float max_dps = 500);
 	void begin();
-	void write(double angle, uint16_t wait_time_ms = 0, double dps = 0);
-	void writeAngle(double angle, uint16_t wait_time_ms = 0, double dps = 0);
+	void write(float angle, uint16_t wait_time_ms = 0, float dps = 0);
+	void writeAngle(float angle, uint16_t wait_time_ms = 0, float dps = 0);
 	void writeMicroseconds(uint16_t pulse_us, uint16_t wait_time_ms = 0);
-	void writeDutyCycle(double duty_cycle);
+	void writeDutyCycle(float duty_cycle);
 	uint16_t getRange() { return _servo.range; };
-	double getMaxDPS() { return _servo.max_dps; };
+	float getMaxDPS() { return _servo.max_dps; };
 
 private:
 	servo_state_t _servo;
@@ -70,7 +70,7 @@ private:
 	static void update(servo_state_t* arg)
 	{
 		uint64_t now = micros();
-		double time_since_last_loop = ((double)now - (double)arg->last_loop) / 1000000;
+		float time_since_last_loop = ((float)now - (float)arg->last_loop) / 1000000;
 		arg->last_loop = now;
 
 		if (motors_enabled())
@@ -82,7 +82,7 @@ private:
 
 				else
 				{
-					double deg_per_loop = arg->dps * time_since_last_loop;
+					float deg_per_loop = arg->dps * time_since_last_loop;
 					if (arg->angle > arg->end_angle)
 					{
 						arg->angle -= deg_per_loop;
@@ -95,11 +95,11 @@ private:
 						if (arg->angle > arg->end_angle) arg->angle = arg->end_angle;
 					}
 
-					double pulse = (double)(arg->angle / arg->range) * (double)(arg->max_pulse_us - arg->min_pulse_us);
+					float pulse = (float)(arg->angle / arg->range) * (float)(arg->max_pulse_us - arg->min_pulse_us);
 					if (arg->servo_dir == DIRECT)
-						pulse = (double)arg->min_pulse_us + pulse;
+						pulse = (float)arg->min_pulse_us + pulse;
 					else
-						pulse = (double)arg->max_pulse_us - pulse;
+						pulse = (float)arg->max_pulse_us - pulse;
 					arg->servo->writeMicroseconds(pulse);
 				}
 			}
